+++ USERS.js
@@ -37,6 +37,14 @@ const users = [
     url: 'https://github.com/zoo-js/zoo',
     logo: 'https://avatars1.githubusercontent.com/u/70757173?s=200&v=4'
   },
+  {
+    url: 'https://github.com/react-component',
+    logo: 'https://avatars3.githubusercontent.com/u/9441414?s=200&v=4'
+  },
+  {
+    url: 'https://github.com/ant-design/pro-components',
+    logo: 'https://avatars1.githubusercontent.com/u/12101536?s=200&v=4'
+  },
 ];
 
 // **************************************************************************